[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.21 ms, Memory: 4.21 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.cpp
@@ -7,7 +7,7 @@ vector<long long> solution(int x, int n) {
     vector<long long> answer(n);
     
     for (int i = 0; i < n; i++) 
-        answer[i] = (long long)x * (i + 1);
+        answer[i] = x * (i + 1);
     
     
     return answer;
